Pin GxEPD2 @ 1.6.9 exact + fix ambiguous Wire1 cast warning
platformio.ini: changed @ >=1.6.9 to @ 1.6.9 (exact). The semver
range syntax isn't reliable with PlatformIO's library index — the
fresh build on Mac Studio pulled an older version. Exact pin ensures
loadImageBuffer is always available.

src/main.cpp: fixed ISO C++ ambiguous overload warning on line 122
by casting CHARGER_ADDR to (uint8_t) in the requestFrom call.

+++ firmware/src/main.cpp
@@ -119,7 +119,7 @@ ChargeState readChargeState() {
   if (Wire1.endTransmission(false) != 0) {
     return CHG_NONE;  // charger not responding = no USB power
   }
-  Wire1.requestFrom(CHARGER_ADDR, (uint8_t)1);
+  Wire1.requestFrom((uint8_t)CHARGER_ADDR, (uint8_t)1);
   if (!Wire1.available()) return CHG_NONE;
 
   uint8_t reg = Wire1.read();
